requirements.txt
requests
beautifulsoup4
streamlit
python-dotenv
langchain
langchain-community
sentence-transformers
pydantic
markdown
numpy
pandas
pypdf
tqdm
google-generativeai
torch
# tensorflow==2.16.1
protobuf
